requirements.txt
Jinja2==3.1.6
python-dotenv==1.1.0
Requests==2.32.3
